ImageRetrievalWithCL/requirements.txt
asttokens==3.0.0
Bottleneck==1.4.2
Brotli==1.0.9
comm==0.2.2
contourpy==1.3.1
cycler==0.11.0
debugpy==1.8.11
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.6.1
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.3
numexpr==2.10.1
numpy==2.1.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.0
PyQt6==6.7.1
PyQt6_sip==13.9.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.2.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
sip==6.10.0
six==1.16.0
stack_data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tornado==6.4.2
traitlets==5.14.3
triton==3.2.0
typing_extensions==4.12.2
tzdata==2023.3
unicodedata2==15.1.0
wcwidth==0.2.13
wheel==0.45.1
zipp==3.21.0
